+++ genmonmaint.sh
@@ -83,6 +83,7 @@ function updatelibraries() {
   sudo $pipcommand install paho-mqtt -U
   sudo $pipcommand install pysnmp -U
   sudo $pipcommand install ldap3 -U
+  sudo $pipcommand install pyasn1 -U
   echo "Done."
 }
 
@@ -138,6 +139,8 @@ function installgenmon() {
     sudo $pipcommand install paho-mqtt
     sudo $pipcommand install pysnmp
     sudo $pipcommand install ldap3
+    # correct problem with LDAP3 module install
+    sudo $pipcommand install pyasn1 -U
 
     sudo chmod 775 "$genmondir/startgenmon.sh"
     sudo chmod 775 "$genmondir/genmonmaint.sh"